% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_fc_vla}
\alias{inrs_fc_vla}
\title{Factor de correccion segun el VLA (metodo INRS)}
\usage{
inrs_fc_vla(vla)
}
\arguments{
\item{vla}{Numeric. VLA en mg/m3.}
}
\value{
Numeric (1, 10, 30 o 100), o \code{NA_real_} si \code{vla} es \code{NA}.
}
\description{
Factor de correccion segun el VLA (metodo INRS)
}
\examples{
inrs_fc_vla(0.05)

}
