% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_liquid_volatility_graph}
\alias{inrs_liquid_volatility_graph}
\title{Volatility class of a liquid from use temperature and boiling point}
\usage{
inrs_liquid_volatility_graph(use_temperature, boiling_point)
}
\arguments{
\item{use_temperature}{Numeric. Use (process) temperature, in degrees
Celsius. Corresponds to the X axis of the graph.}

\item{boiling_point}{Numeric. Boiling point, in degrees Celsius.
Corresponds to the Y axis of the graph.}
}
\value{
Character scalar (\code{"1"} low, \code{"2"} medium, \code{"3"} high).
}
\description{
Classifies the volatility of a liquid by comparing its boiling point
with the two class-separating lines of the INRS method graph
(Figure 2, use temperature on the X axis, boiling point on the Y
axis).
}
\examples{
inrs_liquid_volatility_graph(use_temperature = 20, boiling_point = 200)
inrs_liquid_volatility_graph(use_temperature = 20, boiling_point = 80)

}
