% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_oel_correction_factor}
\alias{inrs_oel_correction_factor}
\title{VLA correction factor (INRS method)}
\usage{
inrs_oel_correction_factor(vla)
}
\arguments{
\item{vla}{Numeric. VLA in mg/m3.}
}
\value{
Numeric (1, 10, 30 or 100), or \code{NA_real_} if \code{vla} is \code{NA}.
}
\description{
VLA correction factor (INRS method)
}
\examples{
inrs_oel_correction_factor(0.05)

}
