% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_potential_exposure_class}
\alias{inrs_potential_exposure_class}
\title{Potential exposure class (INRS method)}
\usage{
inrs_potential_exposure_class(quantity_class, frequency_class)
}
\arguments{
\item{quantity_class}{Character. See \code{\link[=inrs_quantity_class]{inrs_quantity_class()}}.}

\item{frequency_class}{Character. See \code{\link[=inrs_frequency_class]{inrs_frequency_class()}}.}
}
\value{
Character scalar (\code{"0"} to \code{"5"}), or \code{NA_character_} if the
combination is not defined.
}
\description{
Consults Table 4 of the INRS method (quantity class x frequency class)
to obtain the potential exposure class.
}
\examples{
inrs_potential_exposure_class("3", "2")

}
