% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_potential_risk_class}
\alias{inrs_potential_risk_class}
\title{Potential risk class (INRS method)}
\usage{
inrs_potential_risk_class(potential_exposure_class, hazard_class)
}
\arguments{
\item{potential_exposure_class}{Character. See
\code{\link[=inrs_potential_exposure_class]{inrs_potential_exposure_class()}}.}

\item{hazard_class}{Character. See \code{\link[=inrs_hazard_class]{inrs_hazard_class()}}.}
}
\value{
Character scalar (\code{"1"} to \code{"5"}), or \code{NA_character_} if the
combination is not defined.
}
\description{
Consults Table 5 of the INRS method (potential exposure class x hazard
class) to obtain the potential risk class.
}
\examples{
inrs_potential_risk_class("4", "3")

}
