% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_potential_risk_score}
\alias{inrs_potential_risk_score}
\title{Potential risk score (INRS method)}
\usage{
inrs_potential_risk_score(potential_risk_class)
}
\arguments{
\item{potential_risk_class}{Character. See \code{\link[=inrs_potential_risk_class]{inrs_potential_risk_class()}}.}
}
\value{
Numeric (1, 10, 100, 1000 or 10000), or \code{NA_real_}.
}
\description{
Potential risk score (INRS method)
}
\examples{
inrs_potential_risk_score("3")

}
