% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_process_type}
\alias{inrs_process_type}
\title{Process class and score (INRS method)}
\usage{
inrs_process_type(type)
}
\arguments{
\item{type}{Character. One of \code{"Dispersive"}, \code{"Open"},
\code{"Closed/opened regularly"}, \code{"Permanently closed"}.}
}
\value{
A one-row \code{data.frame} with columns \code{class} and \code{score}.
}
\description{
Process class and score (INRS method)
}
\examples{
inrs_process_type("Open")

}
