% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_puntuacion_riesgo_potencial}
\alias{inrs_puntuacion_riesgo_potencial}
\title{Puntuacion de riesgo potencial (metodo INRS)}
\usage{
inrs_puntuacion_riesgo_potencial(clase_riesgo_potencial)
}
\arguments{
\item{clase_riesgo_potencial}{Character. Vease
\code{\link[=inrs_clase_riesgo_potencial]{inrs_clase_riesgo_potencial()}}.}
}
\value{
Numeric (1, 10, 100, 1000 o 10000), o \code{NA_real_}.
}
\description{
Puntuacion de riesgo potencial (metodo INRS)
}
\examples{
inrs_puntuacion_riesgo_potencial("3")

}
