% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_puntuacion_volatilidad}
\alias{inrs_puntuacion_volatilidad}
\title{Puntuacion de volatilidad o pulverulencia (metodo INRS)}
\usage{
inrs_puntuacion_volatilidad(clase_volatilidad)
}
\arguments{
\item{clase_volatilidad}{Character. \code{"1"}, \code{"2"} o \code{"3"}, tal como
devuelven \code{\link[=inrs_volatilidad_liquido_grafico]{inrs_volatilidad_liquido_grafico()}},
\code{\link[=inrs_volatilidad_liquido_presion]{inrs_volatilidad_liquido_presion()}} o \code{\link[=inrs_pulverulencia_solido]{inrs_pulverulencia_solido()}}.}
}
\value{
Numeric (1, 10 o 100), o \code{NA_real_}.
}
\description{
Puntuacion de volatilidad o pulverulencia (metodo INRS)
}
\examples{
inrs_puntuacion_volatilidad("2")

}
