% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_riesgo_inhalacion}
\alias{inrs_riesgo_inhalacion}
\title{Puntuacion final de riesgo por inhalacion (metodo INRS)}
\usage{
inrs_riesgo_inhalacion(
  puntuacion_riesgo_potencial,
  puntuacion_volatilidad,
  puntuacion_procedimiento,
  puntuacion_proteccion,
  fc_vla
)
}
\arguments{
\item{puntuacion_riesgo_potencial}{Numeric. Vease
\code{\link[=inrs_puntuacion_riesgo_potencial]{inrs_puntuacion_riesgo_potencial()}}.}

\item{puntuacion_volatilidad}{Numeric. Vease
\code{\link[=inrs_puntuacion_volatilidad]{inrs_puntuacion_volatilidad()}}.}

\item{puntuacion_procedimiento}{Numeric. Vease \code{\link[=inrs_procedimiento]{inrs_procedimiento()}}.}

\item{puntuacion_proteccion}{Numeric. Vease \code{\link[=inrs_proteccion]{inrs_proteccion()}}.}

\item{fc_vla}{Numeric. Vease \code{\link[=inrs_fc_vla]{inrs_fc_vla()}}.}
}
\value{
Numeric, o \code{NA_real_} si falta algun componente.
}
\description{
Producto de las cinco puntuaciones parciales del metodo INRS.
}
\examples{
inrs_riesgo_inhalacion(100, 10, 0.5, 0.7, 10)

}
