% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_risk_characterisation}
\alias{inrs_risk_characterisation}
\title{Inhalation risk characterisation (INRS method)}
\usage{
inrs_risk_characterisation(inhalation_risk)
}
\arguments{
\item{inhalation_risk}{Numeric. See \code{\link[=inrs_inhalation_risk]{inrs_inhalation_risk()}}.}
}
\value{
Character scalar describing the action priority, or
\code{NA_character_} if \code{inhalation_risk} is \code{NA}.
}
\description{
Inhalation risk characterisation (INRS method)
}
\examples{
inrs_risk_characterisation(2500)

}
