% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_solid_dustiness}
\alias{inrs_solid_dustiness}
\title{Dustiness class of a solid (INRS method)}
\usage{
inrs_solid_dustiness(description)
}
\arguments{
\item{description}{Character. One of \code{"Dust that generates a lot of visible dispersion in the air"}, \code{"Fine dust with little visible dispersion"} or \code{"Compact solid with no visible dust"}.}
}
\value{
Character scalar (\code{"1"} to \code{"3"}), or \code{NA_character_} if
\code{description} does not match any valid option.
}
\description{
Dustiness class of a solid (INRS method)
}
\examples{
inrs_solid_dustiness("Fine dust with little visible dispersion")

}
