% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_volatility_score}
\alias{inrs_volatility_score}
\title{Volatility or dustiness score (INRS method)}
\usage{
inrs_volatility_score(volatility_class)
}
\arguments{
\item{volatility_class}{Character. \code{"1"}, \code{"2"} or \code{"3"}, as returned
by \code{\link[=inrs_liquid_volatility_graph]{inrs_liquid_volatility_graph()}}, \code{\link[=inrs_liquid_volatility_pressure]{inrs_liquid_volatility_pressure()}}
or \code{\link[=inrs_solid_dustiness]{inrs_solid_dustiness()}}.}
}
\value{
Numeric (1, 10 or 100), or \code{NA_real_}.
}
\description{
Volatility or dustiness score (INRS method)
}
\examples{
inrs_volatility_score("2")

}
