% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_coshh}
\alias{run_coshh}
\title{Launch the COSHH Essentials Shiny application}
\usage{
run_coshh(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}} (e.g.
\code{launch.browser}, \code{port}).}
}
\value{
No return value, called for side effects. Launches a Shiny
application in the default browser.
}
\description{
Launch the COSHH Essentials Shiny application
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_coshh()
\dontshow{\}) # examplesIf}
}
