% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_inrs}
\alias{run_inrs}
\title{Launch the INRS method Shiny application}
\usage{
run_inrs(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}} (e.g.
\code{launch.browser}, \code{port}).}
}
\value{
No return value, called for side effects. Launches a Shiny
application in the default browser.
}
\description{
Launch the INRS method Shiny application
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_inrs()
\dontshow{\}) # examplesIf}
}
