% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_une689}
\alias{run_une689}
\title{Launch the UNE-EN 689 Shiny application (preliminary, statistical and
periodic assessment)}
\usage{
run_une689(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}} (e.g.
\code{launch.browser}, \code{port}).}
}
\value{
No return value, called for side effects. Launches a Shiny
application in the default browser.
}
\description{
Launch the UNE-EN 689 Shiny application (preliminary, statistical and
periodic assessment)
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_une689()
\dontshow{\}) # examplesIf}
}
