% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une689_estadistica.R
\name{une689_estadisticos}
\alias{une689_estadisticos}
\title{Estadisticos descriptivos de la exposicion diaria (UNE-EN 689)}
\usage{
une689_estadisticos(ed)
}
\arguments{
\item{ed}{Numeric vector. Valores de exposicion diaria (ED), todos
estrictamente positivos.}
}
\value{
Una lista con los elementos \code{MA}, \code{DS}, \code{MG}, \code{DSG}.
}
\description{
Calcula la media y desviacion tipica aritmeticas (MA, DS) y
geometricas (MG, DSG) de un conjunto de valores de exposicion diaria
(ED), necesarios para contrastar el ajuste normal y lognormal.
}
\examples{
une689_estadisticos(c(5, 6, 7, 8, 9, 10))

}
