% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une689_estadistica.R
\name{une689_periodicidad_opcion1}
\alias{une689_periodicidad_opcion1}
\title{Periodicidad recomendada, opcion 1 (MG o MA frente al VLA)}
\usage{
une689_periodicidad_opcion1(valor_referencia, vla)
}
\arguments{
\item{valor_referencia}{Numeric. MG si la distribucion es lognormal,
o MA si es normal (vease \code{\link[=une689_estadisticos]{une689_estadisticos()}}).}

\item{vla}{Numeric. Valor Limite Ambiental (mg/m3).}
}
\value{
Character escalar describiendo la periodicidad recomendada,
en meses.
}
\description{
Periodicidad recomendada, opcion 1 (MG o MA frente al VLA)
}
\examples{
une689_periodicidad_opcion1(valor_referencia = 0.8, vla = 10)

}
