% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une689_estadistica.R
\name{une689_statistics}
\alias{une689_statistics}
\title{Descriptive statistics of daily exposure (UNE-EN 689)}
\usage{
une689_statistics(ed)
}
\arguments{
\item{ed}{Numeric vector. Daily exposure (ED) values, all strictly
positive.}
}
\value{
A list with elements \code{MA}, \code{DS}, \code{MG}, \code{DSG}.
}
\description{
Calculates the arithmetic mean and standard deviation (MA, DS) and the
geometric mean and standard deviation (MG, DSG) of a set of daily
exposure (ED) values, needed to test the normal and lognormal fits.
}
\examples{
une689_statistics(c(5, 6, 7, 8, 9, 10))

}
