% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_collective_protection}
\alias{inrs_collective_protection}
\title{Clase y puntuacion de proteccion colectiva (metodo INRS)}
\usage{
inrs_collective_protection(situacion)
}
\arguments{
\item{situacion}{Character. Una de las situaciones de la Figura 4
INRS, p. ej. \code{"Captacion envolvente"}.}
}
\value{
Un \code{data.frame} de una fila con columnas \code{clase} y
\code{puntuacion}.
}
\description{
Clase y puntuacion de proteccion colectiva (metodo INRS)
}
\examples{
inrs_collective_protection("Captacion envolvente")

}
