% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_potential_risk_score}
\alias{inrs_potential_risk_score}
\title{Puntuacion de riesgo potencial (metodo INRS)}
\usage{
inrs_potential_risk_score(clase_riesgo_potencial)
}
\arguments{
\item{clase_riesgo_potencial}{Character. Vease
\code{\link[=inrs_potential_risk_class]{inrs_potential_risk_class()}}.}
}
\value{
Numeric (1, 10, 100, 1000 o 10000), o \code{NA_real_}.
}
\description{
Puntuacion de riesgo potencial (metodo INRS)
}
\examples{
inrs_potential_risk_score("3")

}
