% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_process_type}
\alias{inrs_process_type}
\title{Clase y puntuacion de procedimiento (metodo INRS)}
\usage{
inrs_process_type(tipo)
}
\arguments{
\item{tipo}{Character. Uno de \code{"Dispersivo"}, \code{"Abierto"},
\code{"Cerrado/abierto regularmente"}, \code{"Cerrado permanente"}.}
}
\value{
Un \code{data.frame} de una fila con columnas \code{clase} y
\code{puntuacion}.
}
\description{
Clase y puntuacion de procedimiento (metodo INRS)
}
\examples{
inrs_process_type("Abierto")

}
