% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_quantity_class}
\alias{inrs_quantity_class}
\title{Clase de cantidad diaria manipulada (metodo INRS)}
\usage{
inrs_quantity_class(valor, unidad = c("g", "ml", "kg", "l"))
}
\arguments{
\item{valor}{Numeric. Cantidad diaria manipulada.}

\item{unidad}{Character. Una de \code{"g"}, \code{"ml"}, \code{"kg"}, \code{"l"}.}
}
\value{
Character escalar (\code{"1"} a \code{"5"}), o \code{NA_character_} si
\code{valor} es \code{NA}.
}
\description{
Clasifica la cantidad diaria de sustancia manipulada en una de las 5
clases del metodo INRS, segun su unidad.
}
\examples{
inrs_quantity_class(50, "g")
inrs_quantity_class(500, "kg")

}
