% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_risk_characterisation}
\alias{inrs_risk_characterisation}
\title{Caracterizacion del riesgo por inhalacion (metodo INRS)}
\usage{
inrs_risk_characterisation(riesgo_inhalacion)
}
\arguments{
\item{riesgo_inhalacion}{Numeric. Vease \code{\link[=inrs_inhalation_risk]{inrs_inhalation_risk()}}.}
}
\value{
Character escalar describiendo la prioridad de accion, o
\code{NA_character_} si \code{riesgo_inhalacion} es \code{NA}.
}
\description{
Caracterizacion del riesgo por inhalacion (metodo INRS)
}
\examples{
inrs_risk_characterisation(2500)

}
