% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrs.R
\name{inrs_volatility_score}
\alias{inrs_volatility_score}
\title{Puntuacion de volatilidad o pulverulencia (metodo INRS)}
\usage{
inrs_volatility_score(clase_volatilidad)
}
\arguments{
\item{clase_volatilidad}{Character. \code{"1"}, \code{"2"} o \code{"3"}, tal como
devuelven \code{\link[=inrs_liquid_volatility_graph]{inrs_liquid_volatility_graph()}},
\code{\link[=inrs_liquid_volatility_pressure]{inrs_liquid_volatility_pressure()}} o \code{\link[=inrs_solid_dustiness]{inrs_solid_dustiness()}}.}
}
\value{
Numeric (1, 10 o 100), o \code{NA_real_}.
}
\description{
Puntuacion de volatilidad o pulverulencia (metodo INRS)
}
\examples{
inrs_volatility_score("2")

}
