% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_coshh}
\alias{run_coshh}
\title{Lanzar la app Shiny del metodo COSHH Essentials}
\usage{
run_coshh(...)
}
\arguments{
\item{...}{Argumentos adicionales pasados a \code{\link[shiny:runApp]{shiny::runApp()}} (p. ej.
\code{launch.browser}, \code{port}).}
}
\value{
No devuelve nada; lanza la app Shiny de forma bloqueante.
}
\description{
Lanzar la app Shiny del metodo COSHH Essentials
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_coshh()
\dontshow{\}) # examplesIf}
}
