% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_inrs}
\alias{run_inrs}
\title{Lanzar la app Shiny del metodo INRS}
\usage{
run_inrs(...)
}
\arguments{
\item{...}{Argumentos adicionales pasados a \code{\link[shiny:runApp]{shiny::runApp()}} (p. ej.
\code{launch.browser}, \code{port}).}
}
\value{
No return value, called for side effects. Launches a Shiny
application in the default browser.
}
\description{
Lanzar la app Shiny del metodo INRS
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_inrs()
\dontshow{\}) # examplesIf}
}
