% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une689_preliminar.R
\name{une689_exposure_index}
\alias{une689_exposure_index}
\title{Indice de exposicion (IE) de una jornada, segun UNE-EN 689}
\usage{
une689_exposure_index(ed, vla)
}
\arguments{
\item{ed}{Numeric. Exposicion diaria, vease \code{\link[=une689_daily_exposure]{une689_daily_exposure()}}.}

\item{vla}{Numeric. Valor Limite Ambiental (mg/m3).}
}
\value{
Numeric escalar (\code{ed / vla}), o \code{NA_real_} si \code{ed} o \code{vla} no
son validos (\code{vla} debe ser \verb{> 0}).
}
\description{
Indice de exposicion (IE) de una jornada, segun UNE-EN 689
}
\examples{
une689_exposure_index(ed = 9, vla = 10)

}
