% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une689_estadistica.R
\name{une689_lsc}
\alias{une689_lsc}
\title{Limite superior de confianza LSC(95,70) (UNE-EN 689)}
\usage{
une689_lsc(
  tipo,
  ut,
  MA = NA_real_,
  DS = NA_real_,
  MG = NA_real_,
  DSG = NA_real_
)
}
\arguments{
\item{tipo}{Character. \code{"Lognormal"}, \code{"Normal"} o \code{"Ninguna"}, tal
como devuelve \code{\link[=une689_tipo_distribucion]{une689_tipo_distribucion()}}.}

\item{ut}{Numeric. Factor UT, vease \code{\link[=une689_ut]{une689_ut()}}.}

\item{MA, DS}{Numeric. Media y desviacion aritmeticas (solo se usan si
\code{tipo = "Normal"}), vease \code{\link[=une689_estadisticos]{une689_estadisticos()}}.}

\item{MG, DSG}{Numeric. Media y desviacion geometricas (solo se usan
si \code{tipo = "Lognormal"}), vease \code{\link[=une689_estadisticos]{une689_estadisticos()}}.}
}
\value{
Numeric escalar con el LSC(95,70), o \code{NA_real_} si
\code{tipo = "Ninguna"}.
}
\description{
Limite superior de confianza LSC(95,70) (UNE-EN 689)
}
\examples{
une689_lsc("Normal", ut = 2.005, MA = 7.5, DS = 1.87)

}
