% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une689_estadistica.R
\name{une689_ur}
\alias{une689_ur}
\title{Indice de riesgo unilateral UR (UNE-EN 689)}
\usage{
une689_ur(
  tipo,
  vla,
  MA = NA_real_,
  DS = NA_real_,
  MG = NA_real_,
  DSG = NA_real_
)
}
\arguments{
\item{tipo}{Character. \code{"Lognormal"}, \code{"Normal"} o \code{"Ninguna"}.}

\item{vla}{Numeric. Valor Limite Ambiental (mg/m3).}

\item{MA, DS, MG, DSG}{Numeric. Vease \code{\link[=une689_lsc]{une689_lsc()}}.}
}
\value{
Numeric escalar, o \code{NA_real_} si \code{tipo = "Ninguna"}.
}
\description{
Indice de riesgo unilateral UR (UNE-EN 689)
}
\examples{
une689_ur("Normal", vla = 10, MA = 7.5, DS = 1.87)

}
