% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une689_preliminar.R
\name{une689_validate_min_days}
\alias{une689_validate_min_days}
\title{Comprobar el numero minimo de jornadas para la evaluacion preliminar}
\usage{
une689_validate_min_days(n_jornadas, minimo = 3L)
}
\arguments{
\item{n_jornadas}{Integer. Numero de jornadas con datos introducidos.}

\item{minimo}{Integer. Numero minimo exigido (por defecto 3).}
}
\value{
Logical: \code{TRUE} si \code{n_jornadas >= minimo}.
}
\description{
La evaluacion preliminar de UNE-EN 689 exige un minimo de jornadas
evaluadas (habitualmente 3). Funcion de ayuda para validar esto antes
de calcular, tanto desde codigo como desde la futura app Shiny.
}
\examples{
une689_validate_min_days(2)
une689_validate_min_days(3)

}
